Fix undefined symbols in header file of parser
The value types, e.g., std::unique_ptr<...>, are referenced in the
header file of parser, y.tab.c, without inserting the includes specified
in the prologue (the %{ %} block at the top).
This causes error when including the header without linking the
implementation. To fix this, add a "%code requires" block. The codes
inside that block are inserted verbatim to the header file.