Expose data members of nodes as public
This allows visitors to access the data members.
I prefer to treat AST nodes as data classes and centralize operations within
visitors. In this context, exposing data members publicly is a reasonable choice.
However, it's worth noting that in C++, constant-ness doesn't propagate down to
pointers, which can affect the true constancy of objects.
We may revisit the design implications of this decision in the future.