Implement if statement qbe code generation
One important change is to move the '@start' label from BlockStmtNode to
ProgramNode. The reason is that BlockStmtNode cannot output the correct
label to its own block because it doesn't know whether it is a if statement
body or a function. For instance, if we keep '@start' at the BlockStmtNode, then
it will have label '@start' and label '@if.x' at the beginning of a if
statement body. Thus, by moving the part where we create labels
for each block to an upper level, each block can have its correct
starting label.