Rename precedence to resolve confusion
The precedence had the same name as the token, while the former one is
used as a placeholder for the precedence. Therefore, it's better for
them to use different names to avoid possible confusion.
The `%prec IF_WITHOUT_ELSE` simply instructs Bison that the rule `IF
'(' expr ')' stmt` has the same precedence as `IF_WITHOUT_ELSE`.