Resolve include dependency for lex.yy.cpp
Previously, we encountered issues due to the lack of forward declaration
for the lexing function `yylex`. Consequently, we resorted to including
lex.yy.cpp directly in parser.y to access it. The initial problem
stemmed from our inability to insert it into the correct position within
yy.tab.hpp. Recently, I found that the %code directive is aptly suited
for addressing this issue.

For more information, refer to:
https://www.gnu.org/software/bison/manual/html_node/_0025code-Summary.html#index-_0025code-5