Run clang tidy on diff with CI
Added a script for CI execution. This script is not intended for local
user execution due to the following reasons:

1. `clang-tidy-diff` supports exit with a failure code when there are
   errors from version 17 or 18 onward. (The exact version is uncertain.
   For more details, please refer to
   https://github.com/llvm/llvm-project/issues/65000.)
2. `bear` has different command line interfaces between version <= 2.4.x
    and onward. (For more details, please refer to
    https://github.com/rizsotto/Bear?tab=readme-ov-file#how-to-use.)

The script may not function properly if any of these issues arise. In
the CI environment, we ensure that the versions of these dependencies
are up-to-date. Specifically, the version of `clang-tidy-diff` is
specified as `18`, and the version of `Bear` on Ubuntu 22.04 is 3.0.x.