Refactor simple assignment code generation
In the previous assignment expression, we assumed that the left
hand side expression is id expression. However, if we want to
support unary expression on the left hand side, such as `*ptr = 10`,
then we need a way for upper level nodes to know the real number
that represents the specific id. Thus, this commit introduces a
new map that maps register number to id number.