Update `DeclArrNode` and `FuncDefNode` to accept base `Type`
Since the type is held by their base type, `DeclNode`, the type accepted
from the constructor is assigned to `std::unique_ptr<Type>`, which still
erases the static type.
While specifying concrete types in the constructor could restrict the user
from passing an incorrect type, it may also cause the parser to crash.
However, we do not want to enforce this through the grammar, as our
grammar needs to be general enough to handle all language constructs.
We prefer to check the type with the type checker. Therefore, we accept
the base type in the constructor.