Fix expression type of function call & built-in print
Previously, the expression type of a function call defaulted to `int`,
which was derived from our symbol entry system where both the expression
type and parameter types were stored. However, this approach became less
effective when we started splitting the function type. Now that our type
system supports function types, it's more efficient to consolidate the
two types into a single `type`. This commit updates the expression type
of the built-in `print` function to reflect this change. The combination
is not done yet.