Reduce ad-hoc type conversions on functions
By not treating function pointers specially in `GetLLVMType`, most of
the ad-hoc conversions from function types to pointer types are
eliminated. The only remaining conversion is when referencing the ID of
a function; the type of such ID expression should be a pointer type
instead of a function type. Additionally, when using such an ID
expression in a function call, we need to convert it back to a function
type.