Use dynamic cast on reference when should not fail
There is a common pattern in our code where we dynamically cast a
pointer and assert that it is not null. This pattern implies that the
cast should not fail. Therefore, we can use the reference version of
dynamic casting, which throws a `std::bad_cast` exception if the cast
fails.

Additionally, we are using dynamic casting with unique pointers, which
requires a manual `get()` to obtain the underlying pointer. Since
`get()` should only be used when necessary, this change also eliminates
its use.