Check type of enum
With the introduction of enums, the concept of constant expressions must
also be defined.
Ideally, a `ConstExprNode` would wrap the `IdExprNode` representing an
enumeration constant. However, to keep the parser simple without
maintaining symbol tables, it cannot determine at parse time whether an
identifier refers to an enumeration constant.
Additionally, modifying the AST to insert a new node above `IdExprNode`
is incompatible with our visitor pattern.
To address this, `IdExprNode` now holds a `ConstExprNode`, a newly
introduced node currently represented by `IntConstExprNode`.
The type checker determines if an identifier is an enumeration constant
by consulting the symbol table. If so, it constructs a `ConstExprNode`
with the corresponding constant value.